Add voltage axis and offset settings by rotary encoder
Rotary encoder is setting the time axis by default. SW1 and SW2 determine whether to select CH1 or CH2. Click SW1 or SW2 to transition from the time axis setting to the voltage axis setting. Click again to transition to the offset setting. And click again to return to the time axis setting. ; When push SW4 for 3 seconds, Rotate the display 180 degrees; Add UART1 for test;